+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.2.2
+* Added auth methods to help with OAuth and Reader SDK methods:
+    * Use `SquareConnect().getAuthEndpoint(clientId, permissions, locale, session, state)` to get the url for requesting permissions for OAuth.
+    * Use `SquareConnect().createMobileAuthorizationCode(locationId)` to get Mobile Auth Code.
+
 ## 0.2.1
 * Added option for using refresh tokens instead of auth token - consider this beta at this point.
     * To use, call SquareConnect.getInstanceWithRefreshToken(refreshToken, clientId, clientSecret)